# Delayed update Binary Tree for Lazy strategy