… lib/compress/zstd_lazy.h
… lib/compress/zstd_opt.c
… lib/compress/zstd_opt.h
… lib/decompress/zstd_decompress.c
… tests/fuzzer.c