This is a pretty nice speed win.

The new strategy consists in stacking new candidates as if it was a hash chain.
Then, only if there is a need to actually consult the chain, they are batch-updated,
before starting the match search itself.
This is supposed to be beneficial when skipping positions,
which happens a lot when using lazy strategy.

The baseline performance for `btlazy2` on my laptop is :
```
15#calgary.tar       :   3265536 ->    955985 (3.416),  7.06 MB/s , 618.0 MB/s
15#enwik7            :  10000000 ->   3067341 (3.260),  4.65 MB/s , 521.2 MB/s
15#silesia.tar       : 211984896 ->  58095131 (3.649),  6.20 MB/s , 682.4 MB/s
```
(only level 15 remains for `btlazy2`, as this strategy is squeezed between `lazy2` and `btopt`)

After this patch, and keeping all parameters identical,
speed is increased by a pretty good margin (+30-50%),
but compression ratio suffers a bit :
```
15#calgary.tar       :   3265536 ->    958060 (3.408),  9.12 MB/s , 621.1 MB/s
15#enwik7            :  10000000 ->   3078318 (3.249),  6.37 MB/s , 525.1 MB/s
15#silesia.tar       : 211984896 ->  58444111 (3.627),  9.89 MB/s , 680.4 MB/s
```

That's because it keeps `1<<searchLog` as a maximum number of unsorted candidates to update.
For a hash chain, it represents the total number of candidates in the chain,
while for binary tree, it represents the maximum depth of searches.
Keep in mind that a lot of candidates won't even be visited in the binary tree,
since they are filtered out by the sort operation.

As a consequence, in the new implementation,
the effective depth of the binary tree is substantially shorter.

To compensate, it's enough to increase `searchLog` value.
Here is the result after adding just +1 to searchLog :
```
15#calgary.tar       :   3265536 ->    956311 (3.415),  8.32 MB/s , 611.4 MB/s
15#enwik7            :  10000000 ->   3067655 (3.260),  5.43 MB/s , 535.5 MB/s
15#silesia.tar       : 211984896 ->  58113144 (3.648),  8.35 MB/s , 679.3 MB/s
```

aka, almost the same compression ratio as before,
but with a noticeable speed increase (+20-30%).

This modification makes `btlazy2` more competitive.
A new paramgrill round determined it could also be applied to level 13 and 14, so compression level table has been adapted.

Not that this patch will conflict with #964, since they work on same files.
#964 usefulness is less clear though, since it merely moves around speed/ratio trade off,
whereas this patch produces a net gain, albeit on a limited scope (`btlazy2`).