@Cyan4973 2018-01-11
Yes, this is a correct description.